+++ Runtime/Framework/CommandBus/ICommandBus.cs
@@ -1,11 +1,9 @@
 using UnityEngine.Events;
 
 namespace DT.UniStart {
-  public interface ICommand { }
-
   public interface ICommandRepo {
-    UnityAction Add<T>(UnityAction command) where T : ICommand;
-    UnityAction<T> Add<T>(UnityAction<T> command) where T : ICommand;
+    UnityAction Add<T>(UnityAction command);
+    UnityAction<T> Add<T>(UnityAction<T> command);
   }
 
   public interface ICommandBus {
@@ -13,7 +11,7 @@ namespace DT.UniStart {
     /// Push a command to the CommandBus.
     /// If the command is not registered, an exception will be thrown.
     /// </summary>
-    void Push<T>(T arg) where T : ICommand;
+    void Push<T>(T arg);
   }
 
   public static class ICommandBusExtension {
@@ -21,7 +19,7 @@ namespace DT.UniStart {
     /// Push a command to the CommandBus.
     /// If the command is not registered, an exception will be thrown.
     /// </summary>
-    public static void Push<T>(this ICommandBus self) where T : ICommand, new() => self.Push(new T());
+    public static void Push<T>(this ICommandBus self) where T : new() => self.Push(new T());
   }
 
   public interface ICommandCenter : ICommandRepo, ICommandBus { }