+++ Samples~/UniSnake/Scripts/Scene/Play/Commands.cs
@@ -1,8 +1,9 @@
 using DT.UniStart;
 using UnityEngine;
 
-namespace UniSnake.Scene.Play {
-  public record TogglePauseCommand : ICommand;
-  public record MoveSnakeCommand : ICommand;
-  public record SetSnakeDirectionCommand(Vector2Int direction) : ICommand;
+namespace UniSnake.Scene.Play
+{
+  public record TogglePauseCommand;
+  public record MoveSnakeCommand;
+  public record SetSnakeDirectionCommand(Vector2Int direction);
 }